TopBar und Navigation in Komponente ausgelagert
