+++ src/Components/App.tsx
@@ -1,31 +1,17 @@
 import {
-    AppBar,
-    Button,
-    Card,
-    CardHeader,
     Container,
     createStyles,
-    Grid,
     makeStyles,
-    Toolbar,
-    Typography,
 } from '@material-ui/core'
 import CssBaseline from '@material-ui/core/CssBaseline'
 import ThemeProvider from '@material-ui/core/styles/MuiThemeProvider'
-import AddCircleOutlineOutlinedIcon from '@material-ui/icons/AddCircleOutlineOutlined'
-import EditOutlinedIcon from '@material-ui/icons/EditOutlined'
-import HomeOutlinedIcon from '@material-ui/icons/HomeOutlined'
-import InfoOutlinedIcon from '@material-ui/icons/InfoOutlined'
 import React, { FC } from 'react'
-import { BrowserRouter as Router, Link, Route, Switch } from 'react-router-dom'
+import { BrowserRouter as Router } from 'react-router-dom'
 
 import logo from '../icons/logo.svg'
 import { responsiveTheme } from '../theme'
-// Import der Komponenten
-import { Create } from './Create'
-import { Details } from './Details'
-import { Edit } from './Edit'
-import { Home } from './Home'
+import TopBar from './TopBar'
+import Routes from '../Routes/Routes'
 
 const useStyles = makeStyles(theme =>
     createStyles({
@@ -37,30 +23,6 @@ const useStyles = makeStyles(theme =>
     })
 )
 
-const RoutePaths = {
-    HOME: '/',
-    EDIT: '/edit',
-    DETAILS: '/details',
-    CREATE: '/create',
-}
-
-// Anzeigen der Komponenten entsprechend des Pfades
-const Routes = (
-    <Switch>
-        <Route path={RoutePaths.DETAILS}>
-            <Details />
-        </Route>
-        <Route path={RoutePaths.CREATE}>
-            <Create />
-        </Route>
-        <Route path={RoutePaths.EDIT}>
-            <Edit />
-        </Route>
-        <Route path={RoutePaths.HOME}>
-            <Home />
-        </Route>
-    </Switch>
-)
 
 const App: FC = () => {
     const classes = useStyles()
@@ -69,34 +31,11 @@ const App: FC = () => {
         <Router>
             <ThemeProvider theme={responsiveTheme}>
                 <CssBaseline />
-
                 <Container maxWidth="lg">
-                    <AppBar position="fixed">
-                        <Toolbar>
-                            <Typography variant="h6" noWrap>
-                                Projektvorlage
-                            </Typography>
-                            <Link to={RoutePaths.HOME}>
-                                <Button startIcon={<HomeOutlinedIcon />} size="large"></Button>
-                            </Link>
-
-                            <Link to={RoutePaths.DETAILS}>
-                                <Button startIcon={<InfoOutlinedIcon />} size="large"></Button>
-                            </Link>
-
-                            <Link to={RoutePaths.CREATE}>
-                                <Button
-                                    startIcon={<AddCircleOutlineOutlinedIcon />}
-                                    size="large"></Button>
-                            </Link>
-
-                            <Link to={RoutePaths.EDIT}>
-                                <Button startIcon={<EditOutlinedIcon />} size="large"></Button>
-                            </Link>
-                        </Toolbar>
-                    </AppBar>
-
-                    <div className={classes.main}>{Routes}</div>
+                    <TopBar />
+                    <div className={classes.main}>
+                    {Routes}
+                    </div>
                 </Container>
             </ThemeProvider>
         </Router>
